+++ projects/examples/__init__.py
@@ -1,1 +1,1 @@
-"""Some basic examples."""
+"""Some basic examples how to extract information from different project exports."""