+++ projects/examples/read_knxproj.py
@@ -5,39 +5,37 @@ import argparse
 import logging
 from pathlib import Path
 
-from projects.knxproj import KnxprojLoader
+from projects.knxproj import Knxproj
 
 
-def setup_parser() -> argparse.ArgumentParser:
-    """Set up parser."""
+def obtain_knxproj() -> Knxproj:
+    """Set up knxproj."""
 
     description = "Read a knxproj export from the ETS and fiddle around with it."
     parser = argparse.ArgumentParser(description=description)
     parser.add_argument(
         "--knxproj", required=True, type=str, help="Path to the ETS .knxproj export."
     )
-    return parser
+    args = parser.parse_args()
+    knxproj_path = Path(args.knxproj)
+
+    return Knxproj(knxproj_path=knxproj_path)
 
 
 def main():
     """Log all provided group addresses and devices."""
 
-    # Parse arguments
-    parser = setup_parser()
-    args = parser.parse_args()
-    knxproj_path = Path(args.knxproj)
-
-    group_addresses, devices = KnxprojLoader(knxproj_path=knxproj_path).run()
+    knxproj = obtain_knxproj()
 
     logging.info("Group addresses:")
-    for group_addr in group_addresses:
+    for group_addr in knxproj.groupaddresses:
         logging.info("\t%s", group_addr)
 
     logging.info("Devices:")
-    for dev in devices:
+    for dev in knxproj.devices:
         logging.info("\t%s", dev)
 
 
 if __name__ == "__main__":
-    logging.basicConfig(level=logging.DEBUG)
+    logging.basicConfig(level=logging.INFO)
     main()